cloud_run/response_parser/requirements.txt
google-cloud-pubsub==2.13.0
python-dotenv==1.0.0